Add get_version_pins method to utils, remove final VERSION_PINS variable
