+++ src/pyobo/constants.py
@@ -2,9 +2,7 @@
 
 """Constants for PyOBO."""
 
-import json
 import logging
-import os
 import re
 
 import pystow
@@ -13,8 +11,6 @@ __all__ = [
     "RAW_DIRECTORY",
     "DATABASE_DIRECTORY",
     "SPECIES_REMAPPING",
-    "VERSION_PINS",
-    "get_version_pins",
 ]
 
 logger = logging.getLogger(__name__)
@@ -102,43 +98,3 @@ PROVENANCE_PREFIXES = {
     "isbn",
     "issn",
 }
-
-
-def get_version_pins():
-    """Retrieve the resource version pins."""
-    try:
-        version_pins_str = os.getenv("VERSION_PINS")
-        if not version_pins_str:
-            version_pins = {}
-        else:
-            version_pins = json.loads(version_pins_str)
-            invalid_prefixes = []
-            for prefix, version in version_pins.items():
-                if not isinstance(prefix, str) or not isinstance(version, str):
-                    logger.error(
-                        f"The prefix:{prefix} and version:{version} name must both be strings"
-                    )
-                    invalid_prefixes.append(prefix)
-            for prefix in invalid_prefixes:
-                version_pins.pop(prefix)
-    except ValueError as e:
-        logger.error(
-            "The value for the environment variable VERSION_PINS must be a valid JSON string: %s"
-            % e
-        )
-        version_pins = {}
-
-    if version_pins:
-        logger.debug(
-            f"These are the resource versions that are pinned.\n"
-            f"{version_pins}. "
-            f"\nPyobo will download the latest version of a resource if it's "
-            f"not pinned.\nIf you want to use a specific version of a "
-            f"resource, edit your VERSION_PINS environmental "
-            f"variable which is a JSON string to include a prefix and version "
-            f"name."
-        )
-    return version_pins
-
-
-VERSION_PINS = get_version_pins()
